+++ Sources/AsyncHTTPClient/HTTPClient.swift
@@ -81,7 +81,7 @@ public class HTTPClient {
     }
 
     deinit {
-        assert(self.pool.connectionProviderCount == 0)
+        assert(self.pool.count == 0)
         assert(self.state == .shutDown, "Client not shut down before the deinit. Please call client.syncShutdown() when no longer needed.")
     }
